[fix-db] Enhance the Area entity
This commit fixes an issue with saving the retrieved network response to the db.
We separate the entity to model, and introduce a mapper that converts the retrieved data into the entity which is stored to the db.
This change also fixes the unit tests as part of the changes.